feat: Read-aloud highlight moves while reading [PT-186754162]
This enables highlighting of words as they are spoken using the CSS highlighting system available on some newer browsers.  This also requires the browser and host system to support word utterance events.

If highlighting or word utterance events are not supported the existing highlighting of the entire spoken text is maintained.