fix: This is WIP branch to fix the word highlighting issues [PT-186870355]
This removes the parser and instead just tries to find the words uttered using a simple word boundary regex search.

This mostly works but the indexing is off and so sometimes picks previous words in the spoken text.  It shouldn't take much to get this fully working.

If this is worked on again the parser files should also be removed.